requirements.txt
matplotlib
numpy
pandas
Pillow
pytorch-lightning
torch
torchvision
wandb